moved options bar to separate file
